requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlmodel>=0.0.6  # Consider upgrading to the latest stable version
passlib[bcrypt]
python-jose
databases==0.5.2
asyncpg>=0.29.0
python-dotenv
pydantic>=1.10.0  # Make sure to use a compatible version
email-validator
